# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team/members/{userId}:
    patch:
      operationId: adminUpdateMember
      tags: [team]
      summary: Update another member's display name and/or roles (PM only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberAdminInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "403":
          description: Forbidden (not a PM)
        "404":
          description: Member not found
...
components:
  schemas:
...
    Cv:
      type: object
      required: [fileName, contentType, objectPath, sizeBytes, uploadedAt]
      properties:
        fileName:
          type: string
        contentType:
          type: string
        objectPath:
          type: string
        sizeBytes:
          type: integer
        uploadedAt:
          type: string
          format: date-time
...
    Member:
      type: object
      required: [id, email, displayName, roles, joinedAt, lastActivityAt, hasCv]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
        joinedAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        hasCv:
          type: boolean
        cv:
          oneOf:
            - $ref: "#/components/schemas/Cv"
            - type: "null"
...
    MemberAdminInput:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        roles:
          type: array
          items:
            type: string
        clearCv:
          type: boolean
          description: When true, remove the member's CV (PM-only admin action).
...